+++ tokio-io/src/io/mod.rs
@@ -9,6 +9,8 @@
 //! [found online]: https://tokio.rs/docs/getting-started/core/
 //! [low level details]: https://tokio.rs/docs/going-deeper-tokio/core-low-level/
 
+mod async_read;
+mod async_write;
 mod copy;
 mod flush;
 mod read;
@@ -18,6 +20,8 @@ mod read_until;
 mod shutdown;
 mod write_all;
 
+pub use self::async_read::AsyncRead;
+pub use self::async_write::AsyncWrite;
 pub use allow_std::AllowStdIo;
 pub use self::copy::{copy, Copy};
 pub use self::flush::{flush, Flush};
